+++ src/api/load.js
@@ -8,17 +8,18 @@ import {Graph} from "../core/graph";
 import {loadGraph} from "../plugins/github/loadGraph";
 import {TimelineCred} from "../analysis/timeline/timelineCred";
 import {defaultParams, partialParams} from "../analysis/timeline/params";
-import {type PartialTimelineCredParameters} from "../analysis/timeline/params";
+import {type TimelineCredParameters} from "../analysis/timeline/params";
 
 import {type Project} from "../core/project";
 import {setupProjectDirectory} from "../core/project_io";
 import {loadDiscourse} from "../plugins/discourse/loadDiscourse";
 import {type PluginDeclaration} from "../analysis/pluginDeclaration";
 import * as NullUtil from "../util/null";
+import {nodeContractions} from "../plugins/identity/nodeContractions";
 
 export type LoadOptions = {|
   +project: Project,
-  +params: ?PartialTimelineCredParameters,
+  +params: ?$Shape<TimelineCredParameters>,
   +plugins: $ReadOnlyArray<PluginDeclaration>,
   +sourcecredDirectory: string,
   +githubToken: string | null,
@@ -91,7 +92,16 @@ export async function load(
   ]);
 
   const pluginGraphs = await Promise.all(pluginGraphPromises);
-  const graph = Graph.merge(pluginGraphs);
+  let graph = Graph.merge(pluginGraphs);
+  const {identities, discourseServer} = project;
+  if (identities.length) {
+    const serverUrl =
+      discourseServer == null ? null : discourseServer.serverUrl;
+    const contractions = nodeContractions(identities, serverUrl);
+    // Only apply contractions if identities have been specified, since it involves
+    // a full Graph copy
+    graph = graph.contractNodes(contractions);
+  }
 
   const projectDirectory = await setupProjectDirectory(
     project,