+++ src/cli/help.js
@@ -8,6 +8,7 @@ import {help as loadHelp} from "./load";
 import {help as scoresHelp} from "./scores";
 import {help as clearHelp} from "./clear";
 import {help as genProjectHelp} from "./genProject";
+import {help as discourseHelp} from "./discourse";
 
 const help: Command = async (args, std) => {
   if (args.length === 0) {
@@ -21,6 +22,7 @@ const help: Command = async (args, std) => {
     scores: scoresHelp,
     clear: clearHelp,
     "gen-project": genProjectHelp,
+    discourse: discourseHelp,
   };
   if (subHelps[command] !== undefined) {
     return subHelps[command](args.slice(1), std);
@@ -43,6 +45,7 @@ function usage(print: (string) => void): void {
       clear         clear SoucrceCred data
       scores        print SourceCred scores to stdout
       gen-project   print a SourceCred project config to stdout
+      discourse     load a Discourse server into SourceCred
       help          show this help message
 
     Use 'sourcecred help COMMAND' for help about an individual command.