+++ src/cli/sourcecred.js
@@ -10,6 +10,7 @@ import load from "./load";
 import scores from "./scores";
 import clear from "./clear";
 import genProject from "./genProject";
+import discourse from "./discourse";
 
 const sourcecred: Command = async (args, std) => {
   if (args.length === 0) {
@@ -31,6 +32,8 @@ const sourcecred: Command = async (args, std) => {
       return scores(args.slice(1), std);
     case "gen-project":
       return genProject(args.slice(1), std);
+    case "discourse":
+      return discourse(args.slice(1), std);
     default:
       std.err("fatal: unknown command: " + JSON.stringify(args[0]));
       std.err("fatal: run 'sourcecred help' for commands and usage");