+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 
 <!-- Please add new entries just beneath this line. -->
+- Change default alpha from 0.05 to 0.20 (#1391)
+- Enable viewing and changing alpha in the explorer (#1390)
+- Enable combining different user identities together (#1385)
+- Add `sourcecred discourse` for loading Discourse servers (#1374)
 - Breaking: Change output format for the scores command (#1372)
 - Include top nodes for every type in Timeline Cred (#1358)